Move caching tests from test_dispatcher to test_caching
